+++ tests/specs/service.js
@@ -132,7 +132,7 @@
         it('does not affect module when suspended', function() {
 
             var application = new Application();
-            var module = new Module(application);
+            var module = new Module();
             var Service1 = Service.extend({
                 _doActivateModule: function (_module, _application) {
                     assert.fail('It should be desactivated');
@@ -150,7 +150,7 @@
             var module, service, application;
 
             application = new Application();
-            module = new Module(application);
+            module = new Module();
 
             // create a stub service to monitor module activation
             var Service1 = Service.extend({
@@ -168,43 +168,144 @@
         });
     });
 
-    describe('Fossil.Service exposition to module', function () {
-        it('should not be exposed as default', function() {
-            var service, application;
-
-            service = new Service();
-            application = new Application({
-                services: {
-                    'foo': service
-                },
-                modules: {
-                    '': Module
-                }
-            });
-
-            assert.isUndefined(application.getModule('').services.foo);
-        });
-
-        it('should be exposed', function () {
-            var service, application;
+    describe('exposed methods', function () {
+        it('are not required', function() {
+            var service, application, module;
 
             // create a stub service to monitor module activation
             var Service1 = Service.extend({
-                options: {
-                    exposeToModule: true
+                exposedMethods: null
+            });
+            service = new Service1();
+            application = new Application();
+            module = new Module();
+            application.connect('', module);
+            application.use('foo', service);
+        });
+        it('are not exposed to elements by default', function() {
+            var service, application, module;
+            var spy = sinon.spy();
+
+            // create a stub service to monitor module activation
+            var Service1 = Service.extend({
+                exposedMethods: ["methodName"],
+                methodName: spy
+            });
+            service = new Service1();
+            module = new Module();
+            application = new Application();
+            application.connect('', module);
+            application.use('foo', service);
+
+            assert.isUndefined(application.methodName);
+            assert.isUndefined(module.methodName);
+        });
+        it('are exposed to elements', function() {
+            var service, application, module;
+            var spy = sinon.spy();
+
+            // create a stub service to monitor module activation
+            var Service1 = Service.extend({
+                exposedMethods: ["methodName"],
+                methodName: spy,
+                options: { expose: true}
+            });
+            service = new Service1();
+            application = new Application();
+            module = new Module();
+            application.connect('', module);
+            application.use('foo', service);
+
+            assert.isFunction(application.methodName);
+            assert.isFunction(module.methodName);
+
+            application.methodName();
+            assert.ok(spy.calledOnce);
+            module.methodName();
+            assert.ok(spy.calledTwice);
+        });
+
+        it('can be prevented from registration on application', function() {
+            var service, application, module;
+
+            // create a stub service to monitor module activation
+            var Service1 = Service.extend({
+                exposedMethods: ["methodName"],
+                methodName: function () {},
+                options: { expose: true, exposeToApplication: false}
+            });
+            service = new Service1();
+            application = new Application();
+            module = new Module();
+            application.connect('', module);
+            application.use('foo', service);
+
+            assert.isUndefined(application.methodName);
+            assert.isFunction(module.methodName);
+        });
+
+        it('can be prevented from registration on module', function() {
+            var service, application, module;
+
+            // create a stub service to monitor module activation
+            var Service1 = Service.extend({
+                exposedMethods: ["methodName"],
+                methodName: function () {},
+                options: { expose: true, exposeToModule: false}
+            });
+            service = new Service1();
+            application = new Application();
+            module = new Module();
+            application.connect('', module);
+            application.use('foo', service);
+
+            assert.isFunction(application.methodName);
+            assert.isUndefined(module.methodName);
+        });
+    });
+
+    describe('fragments', function () {
+        var Service1, application, module, spy;
+        beforeEach(function() {
+            spy = sinon.spy();
+            Service1 = Service.extend({
+                exposedMethods: ["methodName"],
+                methodName: spy,
+                options: {expose: true},
+                _doActivateFragment: spy
+            });
+            application = new Application({
+                fragments: {
+                    'frag0': Fossil.Fragment
                 }
             });
-
-            application = new Application();
-            service = new Service1();
-
-            application.connect('', Module);
-            application.use('foo', service);
-            application.connect('bar', Module);
-
-            assert.strictEqual(application.getModule('').services.foo, service);
-            assert.strictEqual(application.getModule('bar').services.foo, service);
+            module = new Module({
+                fragments: {
+                    'frag1': Fossil.Fragment
+                }
+            });
         });
+
+        it('should not be processed on load', function() {
+            var service = new Service1();
+            application.use('foo', service);
+            application.connect('', module);
+
+            assert.equal(spy.callCount, 0);
+        });
+        it('should be processed on demand', function() {
+            var service = new Service1();
+            application.use('foo', service);
+            application.connect('', module);
+
+            application.ensureFragment('frag0');
+            assert.ok(spy.calledOnce);
+            module.ensureFragment('frag1');
+            assert.ok(spy.calledTwice);
+        });
+    });
+
+    describe('pass to application', function () {
 
         it('should be possible to define services and modules in options', function (done) {
             this.timeout(15);
@@ -285,6 +386,19 @@
             application.use('service', service);
             assert.ok(service.prefixEvent);
             assert.equal(service.prefixEvent('foo'), 'service:service:foo');
+        });
+    });
+
+    describe('Services can be register during', function () {
+        it('application initialization', function() {
+            var Application1 = Application.extend({
+                initialize: function () {
+                    this.use('foo', new Service());
+                }
+            });
+
+            var app = new Application1();
+            assert.ok(app.services.foo);
         });
     });
 })(chai.assert, sinon, _, Fossil.Service, Fossil.Application, Fossil.Module);