+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 dev
 ---
 
+* Services `expose` config option exposes methods. Service itself is always
+  exposed on service property.
 * `Fossil.Mixins.Layoutable`: Introduce `setLayout`, a method to set/replace layout.
 * `Fossil.Mixins.Fragmentables`: accepts `fragments` option.